LayoutTests/imported/w3c/web-platform-tests/notifications/permissions-non-secure-expected.txt


PASS Notification.requestPermission must be called from a secure context
FAIL Notification.permission must be called from a secure context assert_equals: expected "denied" but got "granted"
